Added Hash Function for Client side Usage
Right now the code should be fully functional to get phi0 and phi1 but haven't done tests for correctness.
The code can right now be seen on src/client/main.c and main functionality is on src/core/utils.c but
can always be switched. It also needs better comments for cross referencing with the paper and from
the documentation